% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indicies.R
\name{calculate_indices}
\alias{calculate_indices}
\title{Calculate Indices Based on RGB and NIR Raster Layers}
\usage{
calculate_indices(red, green, blue, nir, indices, output_dir = NULL)
}
\arguments{
\item{red}{SpatRaster representing the Red channel.}

\item{green}{SpatRaster representing the Green channel.}

\item{blue}{SpatRaster representing the Blue channel.}

\item{nir}{SpatRaster representing the Near-Infrared channel.}

\item{indices}{Character vector specifying which indices to calculate.
Options include "NDVI", "EVI", "SAVI", "GCI", "OSAVI", "MSAVI", "TVI",
"DVI", "RVI", "IPVI", "GNDVI", "VARI", "TGI", "NDRE", "CIgreen",
"CIrededge", "ARVI", "MTVI", "NDWI", "NDSI", "BSI", "NBRI", "IBI".}

\item{output_dir}{Character string specifying the directory to save the output indices as .tif files.
If NULL, the function returns a SpatRaster stack.}
}
\value{
A SpatRaster containing the calculated indices if output_dir is NULL, otherwise NULL.
}
\description{
This function calculates various indices based on the provided RGB and NIR raster layers.
Indices include NDVI, EVI, SAVI, GCI, and many others.
}
